+++ output.lua
@@ -24,7 +24,7 @@ local function macroReplace(input, macro_table)
     return input
 end
 
-function BuffDuty.printDuties(cmd, channel_type, duty_table)
+function BuffDuty.printDuties(cmd, channel_type, channel_id, duty_table)
     if not duty_table then return end
 
     -- Local aliases
@@ -51,7 +51,7 @@ function BuffDuty.printDuties(cmd, channel_type, duty_table)
     -- Only a single player has been assigned buffing duty
     if duty_count == 1 then
         local player_name, player_macros = next(duty_table)
-        if channel_type == BuffDuty.WHISPER_CHANNEL_TYPE then
+        if channel_type == BuffDuty.CHANNELS.WHISPER then
             local single_whisper = cmd.single_whisper or messages.single_whisper
             player_macros["class"] = duty_macros["class"] -- Add the "class" to the player macros
             single_whisper = macroReplace(single_whisper, player_macros)
@@ -59,19 +59,19 @@ function BuffDuty.printDuties(cmd, channel_type, duty_table)
                 printInfoMessage(single_whisper, "ffff78ea")
             else
                 single_whisper = string.format(messages.message_title, single_whisper)
-                SendChatMessage(single_whisper, BuffDuty.WHISPER_CHANNEL_TYPE, nil, player_name)
+                SendChatMessage(single_whisper, channel_type, nil, player_name)
             end
         else -- Public channel
             local single_message = string.format(messages.message_title, cmd.single_message or messages.single_message)
             duty_macros["name"] = player_name -- Add the player "name" to the duty info
             single_message = macroReplace(single_message, duty_macros)
-            SendChatMessage(single_message, channel_type, nil, cmd.channel_name)
+            SendChatMessage(single_message, channel_type, nil, channel_id)
         end
         return
     end
 
     -- Multiple players have been assinged buffing duty
-    if channel_type == BuffDuty.WHISPER_CHANNEL_TYPE then 
+    if channel_type == BuffDuty.CHANNELS.WHISPER then
         local whisper_message = cmd.duty_whisper or messages.duty_whisper
         for player_name, player_macros in pairs(duty_table) do -- For each player
             local duty_whisper = macroReplace(whisper_message, player_macros)
@@ -79,18 +79,18 @@ function BuffDuty.printDuties(cmd, channel_type, duty_table)
                 printInfoMessage(duty_whisper, "ffff78ea")
             else
                 whisper_message = string.format(messages.message_title, whisper_message)
-                SendChatMessage(duty_whisper, BuffDuty.WHISPER_CHANNEL_TYPE, nil, player_name)
+                SendChatMessage(duty_whisper, channel_type, nil, player_name)
             end
         end
     else -- Public channel
         local public_title = string.format(messages.message_title, cmd.public_title or messages.public_title)
         public_title = macroReplace(public_title, duty_macros)
-        SendChatMessage(public_title, channel_type, nil, cmd.channel_name)
+        SendChatMessage(public_title, channel_type, nil, channel_id)
         -- Duty lines
         for player_name, player_macros in pairs(duty_table) do
             local duty_line = cmd.duty_line or messages.duty_line
             duty_line = macroReplace(duty_line, player_macros)
-            SendChatMessage(duty_line, channel_type, nil, cmd.channel_name)
+            SendChatMessage(duty_line, channel_type, nil, channel_id)
         end
     end
 end