+++ utils.lua
@@ -1,12 +1,29 @@
 local Utils = {}
 BuffDuty.Utils = Utils
 
-function Utils.tableContainsValue(table, val)
-    if not table or not val then
+function Utils.containsStringValue(table, string_value)
+    if not table or not string_value then
         return false
     end
+    string_value = string_value:lower()
     for _, value in pairs(table) do
-        if value:lower() == val:lower() then
+        if string_value == string.lower(value) then
+            return true
+        end
+    end
+    return false
+end
+
+function Utils.containsName(table, name, get_name_func)
+    if not table or not name then
+        return false
+    end
+    if not get_name_func then
+        get_name_func = function(x) return x end
+    end
+    name = name:lower()
+    for _, value in pairs(table) do
+        if name == string.lower(get_name_func(value)) then
             return true
         end
     end
@@ -43,12 +60,31 @@ function Utils.sortStringArray(string_array)
     end)
 end
 
-function Utils.stringSplit(input, seperator)
+function Utils.stringSplit(input, separator)
     local list = {}
-    for value in string.gmatch(input, "([^"..seperator.."]+)") do -- Match all characters between seperators
-        table.insert(list, value)
+    local idx = 0
+    for value in string.gmatch(input, "([^"..separator.."]+)") do -- Match all characters between separators
+        idx = idx + 1
+        list[idx] = value
     end
     return list
+end
+
+function Utils.stringSplitAsFlags(input, separator)
+    local list = {}
+    for value in string.gmatch(input, "([^"..separator.."]+)") do -- Match all characters between separators
+        if value and #value > 1 then
+            list[value:lower()] = true
+        end
+    end
+    return list
+end
+
+function Utils.stringTrim(value, start_patten, end_patten)
+    start_patten = start_patten or "%s*" -- Default to whitespace characters
+    end_patten = end_patten or "%s*"
+    value = string.gsub(value, "^"..start_patten.."(.-)"..end_patten.."$", "%1")
+    return value
 end
 
 function Utils.stringTitleCase(input)